Move the collapsible component into the "IncomeNotifications" component
