+++ frontend/src/employee-frontend/components/person-profile/PersonIncome.tsx
@@ -14,7 +14,6 @@ import type { IncomeId, PersonId } from 'lib-common/generated/api-types/shared'
 import { useMutationResult, useQueryResult } from 'lib-common/query'
 import Pagination from 'lib-components/Pagination'
 import { AddButtonRow } from 'lib-components/atoms/buttons/AddButton'
-import { CollapsibleContentArea } from 'lib-components/layout/Container'
 import { FixedSpaceRow } from 'lib-components/layout/flex-helpers'
 import { PersonName } from 'lib-components/molecules/PersonNames'
 import { H3 } from 'lib-components/typography'
@@ -46,15 +45,9 @@ interface Props {
   id: PersonId
 }
 
-const StyledCollapsibleContentArea = styled(CollapsibleContentArea)`
-  width: 50%;
-`
-
 export default React.memo(function PersonIncome({ id }: Props) {
   const { i18n } = useTranslation()
   const { permittedActions } = useContext(PersonContext)
-
-  const [open, setOpen] = useState(false)
 
   const children = useQueryResult(
     incomeStatementChildrenQuery({ guardianId: id })
@@ -65,21 +58,7 @@ export default React.memo(function PersonIncome({ id }: Props) {
       <H3>{i18n.personProfile.incomeStatement.title}</H3>
       <IncomeStatements personId={id} />
       <Gap size="L" />
-      <StyledCollapsibleContentArea
-        data-qa="income-notifications-collapsible"
-        aria-label={i18n.common.showMore}
-        paddingHorizontal="zero"
-        opaque
-        open={open}
-        toggleOpen={() => setOpen(!open)}
-        title={
-          <H3 noMargin>
-            {i18n.personProfile.incomeStatement.notificationsTitle}
-          </H3>
-        }
-      >
-        <IncomeNotifications personId={id} />
-      </StyledCollapsibleContentArea>
+      <IncomeNotifications personId={id} />
       <Gap size="L" />
       {renderResult(children, (children) => (
         <>
